Added the missing "?context=1" to the DeltaLog
This feature was missing from the port over from `delta-bot-three` and I've noticed the difference in my usage of /r/DeltaLog. It's annoying because without this, you have to do two clicks to actually read what the person originally posted that caused the !delta to be given! All I did was concat "?context=1" just like what was done in the `delta-bot-three` implementation. 

You can see the original code as it was done in the old repo here: https://github.com/MystK/delta-bot-three/blob/master/i18n/index.json#L20